# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TLC platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /session/login:
    post:
      operationId: login
      tags: [session]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
...
  /trainer/participants/{id}:
    get:
      operationId: getTrainerParticipant
      tags: [trainer]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Trainer participant detail (or null)
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: "#/components/schemas/TrainerParticipantDetail" }
                  - { type: "null" }
  /trainer/events:
    get:
      operationId: getTrainerEventsData
      tags: [trainer]
      responses:
        "200":
          description: Trainer events data
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrainerEventsData" }
    post:
      operationId: createEvent
      tags: [trainer]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateEventRequest" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IdResult" }
  /trainer/resources:
    get:
      operationId: getTrainerResourcesData
      tags: [trainer]
      responses:
        "200":
          description: Trainer resources data
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrainerResourcesData" }
    post:
      operationId: createResource
      tags: [trainer]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateResourceRequest" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IdResult" }
  /trainer/resources/{id}/status:
    post:
      operationId: setResourceStatus
      tags: [trainer]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetResourceStatusRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /trainer/enrollments/{id}/certificate:
    post:
      operationId: issueCertificate
      tags: [trainer]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Certificate result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CertificateResult" }

  # ------------------------------------------------------------------- admin
...
components:
  schemas:
...
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...
    IdResult:
      type: object
      required: [ok, id]
      properties:
        ok: { type: boolean }
        id: { type: string }
    Role:
      type: string
      enum: [PARTICIPANT, COMPANY_VIEWER, TRAINER, ADMIN, SUPER_ADMIN]
    SessionUser:
      type: object
      required: [id, role, companyId, name, email]
      properties:
        id: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        companyId: { type: string, nullable: true }
        name: { type: string, nullable: true }
        email: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    LoginResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/SessionUser" }
    CompanyLite:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
...
    UserLite:
      type: object
      required: [id, email]
      properties:
        id: { type: string }
        name: { type: string, nullable: true }
        email: { type: string }
        image: { type: string, nullable: true }
        title: { type: string, nullable: true }
    Pillar:
      type: string
      enum: [EQ, IQ, MQ]
...
    ModuleLite:
      type: object
      required: [id, title, pillar]
      properties:
        id: { type: string }
        title: { type: string }
        pillar: { $ref: "#/components/schemas/Pillar" }
        order: { type: integer }
        weekNo: { type: integer, nullable: true }
        summary: { type: string, nullable: true }
        programId: { type: string }
    EventLite:
      type: object
      required: [id, cohortId, type, title, startAt, endAt]
      properties:
        id: { type: string }
        cohortId: { type: string }
        moduleId: { type: string, nullable: true }
        type: { type: string }
        title: { type: string }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        joinUrl: { type: string, nullable: true }
        location: { type: string, nullable: true }
        status: { type: string }
        weekNo: { type: integer, nullable: true }
...
    ModuleProgressItem:
      type: object
      required: [id, enrollmentId, weekNo, status]
      properties:
        id: { type: string }
        enrollmentId: { type: string }
        weekNo: { type: integer }
        moduleId: { type: string, nullable: true }
        status: { type: string }
        completedAt: { type: string, format: date-time, nullable: true }
        module:
          oneOf: [{ $ref: "#/components/schemas/ModuleLite" }, { type: "null" }]
...
    CertificateLite:
      type: object
      required: [id, serial, issuedAt]
      properties:
        id: { type: string }
        serial: { type: string }
        issuedAt: { type: string, format: date-time }
...
    ResourceItem:
      type: object
      required: [id, title, type, status, printReady]
      properties:
        id: { type: string }
        title: { type: string }
        type: { type: string }
        status: { type: string }
        fileKey: { type: string, nullable: true }
        url: { type: string, nullable: true }
        description: { type: string, nullable: true }
        printReady: { type: boolean }
        moduleId: { type: string, nullable: true }
        cohortId: { type: string, nullable: true }
        programId: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        module:
          oneOf: [{ $ref: "#/components/schemas/ModuleLite" }, { type: "null" }]
...
    CohortLite:
      type: object
      required: [id, name, status]
      properties:
        id: { type: string }
        name: { type: string }
        status: { type: string }
        sessionDay: { type: string, nullable: true }
        startDate: { type: string, format: date-time }
        isPrivate: { type: boolean }
        capacity: { type: integer }
...
    TrainerParticipantDetail:
      type: object
      required: [id, status, user, cohort, moduleProgress]
      properties:
        id: { type: string }
        status: { type: string }
        user: { $ref: "#/components/schemas/UserLite" }
        companyName: { type: string, nullable: true }
        cohort: { $ref: "#/components/schemas/CohortLite" }
        moduleProgress:
          type: array
          items: { $ref: "#/components/schemas/ModuleProgressItem" }
        certificate:
          oneOf: [{ $ref: "#/components/schemas/CertificateLite" }, { type: "null" }]
    EventWithCohort:
      allOf:
        - $ref: "#/components/schemas/EventLite"
        - type: object
          properties:
            cohortName: { type: string }
    ModuleOption:
      type: object
      required: [id, title]
      properties:
        id: { type: string }
        title: { type: string }
        pillar: { type: string }
        weekNo: { type: integer, nullable: true }
    TrainerEventsData:
      type: object
      required: [cohorts, events, modules]
      properties:
        cohorts:
          type: array
          items: { $ref: "#/components/schemas/CompanyLite" }
        events:
          type: array
          items: { $ref: "#/components/schemas/EventWithCohort" }
        modules:
          type: array
          items: { $ref: "#/components/schemas/ModuleOption" }
    ResourceWithCohort:
      allOf:
        - $ref: "#/components/schemas/ResourceItem"
        - type: object
          properties:
            cohortName: { type: string, nullable: true }
    TrainerResourcesData:
      type: object
      required: [cohorts, resources, modules]
      properties:
        cohorts:
          type: array
          items: { $ref: "#/components/schemas/CompanyLite" }
        resources:
          type: array
          items: { $ref: "#/components/schemas/ResourceWithCohort" }
        modules:
          type: array
          items: { $ref: "#/components/schemas/ModuleOption" }
    CreateEventRequest:
      type: object
      required: [cohortId, type, title, startAt, endAt]
      properties:
        cohortId: { type: string }
        type: { type: string }
        title: { type: string }
        startAt: { type: string }
        endAt: { type: string }
        joinUrl: { type: string }
        location: { type: string }
        weekNo: { type: integer, nullable: true }
        moduleId: { type: string }
    CreateResourceRequest:
      type: object
      required: [cohortId, title, type]
      properties:
        cohortId: { type: string }
        title: { type: string }
        type: { type: string }
        moduleId: { type: string, nullable: true }
        fileKey: { type: string, nullable: true }
        printReady: { type: boolean }
        description: { type: string, nullable: true }
    SetResourceStatusRequest:
      type: object
      required: [status]
      properties:
        status: { type: string }
    CertificateResult:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
        serial: { type: string }
        error: { type: string }
...