# lib/api-zod/src/generated/types/resourceItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * TLC platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ModuleLite } from './moduleLite';

export interface ResourceItem {
  id: string;
  title: string;
  type: string;
  status: string;
  fileKey?: string | null;
  url?: string | null;
  description?: string | null;
  printReady: boolean;
  moduleId?: string | null;
  cohortId?: string | null;
  programId?: string | null;
  createdAt?: Date;
  module?: ModuleLite | null;
}