# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * TLC platform API
 * OpenAPI spec version: 0.1.0
 */

export type Role = typeof Role[keyof typeof Role];


export const Role = {
  PARTICIPANT: 'PARTICIPANT',
  COMPANY_VIEWER: 'COMPANY_VIEWER',
  TRAINER: 'TRAINER',
  ADMIN: 'ADMIN',
  SUPER_ADMIN: 'SUPER_ADMIN',
} as const;