# lib/api-zod/src/generated/types/trainerResourcesData.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * TLC platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CompanyLite } from './companyLite';
import type { ModuleOption } from './moduleOption';
import type { ResourceWithCohort } from './resourceWithCohort';

export interface TrainerResourcesData {
  cohorts: CompanyLite[];
  resources: ResourceWithCohort[];
  modules: ModuleOption[];
}